Messages script debugged and workjin

+++ addmessage.php
@@ -1,12 +1,29 @@
 <?php
-session_start();
-if ($_SESSION['check'] != TRUE){ 
-	if ($_POST['password'] ){
-		require 'password.php';
-		if ($password == sha1($_POST['password'])){
-			$_SESSION['check'] = TRUE;
-		}
-	} else { ?>
+
+if(session_status() != PHP_SESSION_ACTIVE) {
+	session_start();
+}
+
+if (!isset($_SESSION['check'])) {$_SESSION['check'] = false;}
+
+require 'password.php';
+include 'connection.php';
+
+
+if ($_SESSION['check'] == false) { 
+	if (isset($_POST['password'])){
+		
+		if (password_verify($_POST["password"], $password)){
+			$_SESSION['check'] = true;
+			
+		} 
+	} 
+}
+
+
+if (!$_SESSION["check"]) {
+
+	echo <<<EOF
 	<!DOCTYPE html>
 	<html>
 	<head>
@@ -26,13 +43,12 @@ if ($_SESSION['check'] != TRUE){
 		</form>
 	</body>
 	</html>
-<?php
-die();
+EOF;
+	die();
 }
-}
-include 'connection.php';
-$con = getConnection();
-if ($_POST){
+
+
+if (isset($_POST["messagesubmit"])){
 	$sql = "INSERT INTO `status_messages` (entryDate, expirationDate, heading, body) VALUES (STR_TO_DATE('" . $_POST['entryDate'] . " " . $_POST['entryTime'] . "', '%m/%d/%Y %H:%i'), STR_TO_DATE('" . $_POST['expirationDate'] . " " . $_POST['expirationTime'] . "', '%m/%d/%Y %H:%i'),  '" . $_POST['heading'] . "', '" . $_POST['body'] . "')";
 	if ($con->query($sql)){
 		$m = "Message added successfully.";
@@ -42,7 +58,7 @@ if ($_POST){
 		$e = TRUE;
 	}
 }
-if (isset($_GET['delete'])){
+if (!isset($_POST["messagesubmit"]) && isset($_GET['delete'])){
 	$sql = "DELETE FROM `status_messages` WHERE messageId = " . $_GET['delete'];
 	if ($con->query($sql)){
 		$m = "Message deleted successfully";
@@ -54,7 +70,7 @@ if (isset($_GET['delete'])){
 }
 $sql = "SELECT * FROM `status_messages` WHERE entryDate < NOW() AND NOW() < expirationDate";
 $res = $con->query($sql);
-if ($res){
+if ($res && $res->num_rows > 0){
 	$messages = $res->fetch_assoc();
 }
 ?>
@@ -72,7 +88,7 @@ if ($res){
 </head>
 <body>
 	<h1>Current Message</h1>
-	<?php if ($messages){ ?>
+	<?php if (isset($messages)){ ?>
 
 		<table cellpadding="5">
 			<thead>
@@ -98,22 +114,22 @@ if ($res){
 	<h2>No current message. Why not add a new one?</h2>
 <?php } ?>
 <h1>Add Status Message</h1>
-<?php if ($m){?>
+<?php if (isset($m)){?>
 <h2 style="color: <?php echo $e ? 'red' : 'darkgreen'; ?>"><?php echo $m;?></h2>
 <?php } ?>
 
 <form method="post">
 	<label for="expirationDate">Entry Date/Time</label><br>
-	<input type="text" class="date" name="entryDate" id="entryDate" size="25">
-	<input type="text" class="time" name="entryTime" id="entryTime" size="10"><br>
+	<input type="text" class="date" name="entryDate" id="entryDate" size="25" required>
+	<input type="text" class="time" name="entryTime" id="entryTime" size="10" required><br>
 	<label for="expirationDate">Expiration Date/Time</label><br>
-	<input type="text" class="date" name="expirationDate" id="expirationDate" size="25">
-	<input class="time" type="text" name="expirationTime" id="expirationTime" size="10"><br>
+	<input type="text" class="date" name="expirationDate" id="expirationDate" size="25" required>
+	<input class="time" type="text" name="expirationTime" id="expirationTime" size="10" required><br>
 	<label for="heading">Heading</label><br>
-	<input type="text" size="50" maxlength="255"name="heading"><br>
-	<label for="body">Expiration Date/Time</label><br>
-	<textarea rows="4" cols="50"name="body"></textarea><br>
-	<input type="submit" value="Submit">
+	<input type="text" size="50" maxlength="255"name="heading" required><br>
+	<label for="body">Message</label><br>
+	<textarea rows="4" cols="50"name="body" required></textarea><br>
+	<input type="submit" name="messagesubmit" value="Submit">
 </form>
 <script src="js/jquery-1.11.1.min.js"></script>
 <script src="js/jquery.timepicker.js"></script>
